facttic/mv-issues#12 refactor and finish post tests

+++ post/dao.js
@@ -5,13 +5,11 @@ const mongooseDelete = require("mongoose-delete");
 const { PostSchema } = require("./model");
 
 PostSchema.statics.createNew = async function createNew(post) {
-  const _post = new PostDAO(post);
-  const newPost = await _post.save();
-  return newPost;
+  return await PostDAO.create(post);
 };
 
 PostSchema.statics.createMany = async function createMany(posts) {
-  return await PostDAO.createMany(posts, { ordered: false });
+  return await PostDAO.insertMany(posts, { ordered: false });
 };
 
 PostSchema.statics.getAll = async function getAll({ skip, limit, sort, query }) {
@@ -47,7 +45,10 @@ PostSchema.statics.getAllByManifestationId = async function getAllByManifestatio
   };
 };
 
-PostSchema.statics.getByPostIdStr = async function getByPostIdStr(postIdStr, source) {
+PostSchema.statics.getByPostIdStrBySource = async function getByPostIdStrBySource(
+  postIdStr,
+  source,
+) {
   return await PostDAO.findOne({ post_id_str: postIdStr, source }).exec();
 };
 
@@ -55,7 +56,11 @@ PostSchema.statics.getById = async function getById(_id) {
   return await PostDAO.findOne({ _id });
 };
 
-PostSchema.statics.removeByUserId = async function removeById(userIdStr, userId) {
+PostSchema.statics.removeByUserIdStr = async function removeByUserIdStr(userIdStr, userId) {
+  const post = await PostDAO.findOne({ "user.id_str": userIdStr });
+  if (!post) {
+    throw new Error("Post does not exist");
+  }
   return await PostDAO.delete({ "user.id_str": userIdStr }, userId);
 };
 
@@ -72,6 +77,10 @@ PostSchema.statics.removeByManifestationId = async function removeByManifestatio
   _id,
   userId = null,
 ) {
+  const post = await PostDAO.findOne({ _id, manifestation_id: manifestationId });
+  if (!post) {
+    throw new Error("Post does not exist");
+  }
   return await PostDAO.delete({ _id, manifestation_id: manifestationId }, userId);
 };
